+++ src/1_Page/PostEdit/model/type.d.ts
@@ -1,0 +1,3 @@
+type UseHookFormProps = {
+  boadDetail: BoardDetails;
+};
