+++ src/1_Page/PostEdit/type.d.ts
@@ -1,0 +1,6 @@
+type FormFields = {
+  category: number;
+  board_list_title: string;
+  board_list_content: string;
+  board_list_img?: FileList;
+};
