+++ src/1_Page/index.tsx
@@ -51,7 +51,10 @@ const Page = () => {
         <Route path={`${PAGE_URI.SIGNUP}`} element={<SignUp />} />
         <Route path={`${PAGE_URI.OAUTH}`} element={<OAuthHub />} />
         <Route path={`${PAGE_URI.POST}/:postId`} element={<PostDetail />} />
-        <Route path={`${PAGE_URI.POST}/write/:postId`} element={<PostEdit />} />
+        <Route
+          path={`${PAGE_URI.POST}/:category/:postId`}
+          element={<PostEdit />}
+        />
       </Routes>
     </div>
   );
