+++ src/3_Entity/Board/useGetBoardDetail.ts
@@ -10,6 +10,7 @@ const useGetBoardDetail = (board_list_idx: number): [BoardDetails, boolean] => {
   const navigate = useNavigate();
 
   React.useEffect(() => {
+    if (board_list_idx < 0) return;
     const endpoint = `/board/${board_list_idx}`;
     request("GET", endpoint, null, true);
   }, []);
@@ -27,7 +28,6 @@ const useGetBoardDetail = (board_list_idx: number): [BoardDetails, boolean] => {
     switch (serverState.status) {
       case 200:
         return;
-
       default:
         alert("게시글을 불러오는 데 실패했습니다.");
         navigate(`/404`);
